changed the way a member uploads documents

+++ app/Http/Controllers/DocumentsController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Member;
 use App\Document;
 use Illuminate\Support\Facades\Storage;
+use App\Http\Middleware\applicationStatus;
 
 class DocumentsController extends Controller
 {
@@ -15,46 +16,41 @@ class DocumentsController extends Controller
               'document' => 'required',
               'documentType'=>'required'
               ]);
-
       $member = Member::find(auth()->user()->id);
+      $docTypes = DocumentsController::docTypes();
 
       if(count($member->document) <= 15){
         //member only allowed to store maximum 15 files, else do not store
         if(request()->hasFile(request()->file)){
           if(request()->file('document')->isvalid()){
             $fileExtension = request()->document->getClientOriginalExtension();
-            
-              $uploadedDocs = [];
-              foreach($member->document->all() as $document)
-                array_push($uploadedDocs, $document->type);
+            $uploadedDocs = [];
+            foreach($member->document->all() as $document)
+              array_push($uploadedDocs, $document->type);
             
             switch (request()->documentType) {
-              case 'application':
-                if(in_array('application',$uploadedDocs))
-                  return redirect('profile')->withErrors(['Application form already uploaded!']);
-                $documentType = "application";
-                $fileNameToStore = "";
-                $member->status = "verifyingApplication";
-                $member->save();
+              case $docTypes['ap']:
+                if(in_array($docTypes['ap'],$uploadedDocs))
+                  return redirect()->back()->withErrors(['Application form already uploaded!']);
+                $documentType = $docTypes['ap'];
+                //$member->miscs->status = "verifyingApplication";
+                //$member->save();
                 break;
-              case 'idpassport':
-                $documentType = "idpassport";
-                $fileNameToStore = "";
-                if(in_array('idpassport',$uploadedDocs))
-                  return redirect('profile')->withErrors(['ID or Passport already uploaded!']);
+              case $docTypes['id']:
+                $documentType = $docTypes['id'];
+                if(in_array($docTypes['id'],$uploadedDocs))
+                  return redirect()->back()->withErrors(['ID or Passport already uploaded!']);
                 break;
-              case 'proofop':
-                if(in_array('proofop',$uploadedDocs))
-                  return redirect('profile')->withErrors(['Proof of Payment already uploaded!']);
-                $documentType = "proofop";
-                $fileNameToStore = "";
+              case $docTypes['pr']:
+                if(in_array($docTypes['pr'],$uploadedDocs))
+                  return redirect()->back()->withErrors(['Proof of Payment already uploaded!']);
+                $documentType = $docTypes['pr'];
                 break;
-              case 'supportingdoc':
-                $documentType = "supportingdoc";
-                $fileNameToStore = "";
+              case $docTypes['su']:
+                $documentType = $docTypes['su'];
                 break;
               default:
-                return redirect('profile')->withErrors(['Something went wrong! Or You are being very naughty!']);
+                return redirect()->back()->withErrors(['Something went wrong! Or You are being very naughty!']);
                 $documentType = "illegalfile";
                 $fileNameToStore = "illegalfile";
                 break;
@@ -65,12 +61,13 @@ class DocumentsController extends Controller
             $fileNameToStore = $member->id."D".$document->id.$documentType.".".$fileExtension;
             $path = "documents";
             $document->type = $documentType;
+            $document->original_name = $request->document->getClientOriginalName();
             $document->extension = $fileExtension;
             $document->path = "/".$path."/".$fileNameToStore;
             $document->save();
             request()->document->storeAs($path,$fileNameToStore);
 
-            return redirect('profile')->withErrors(['success','File upload successful']);
+            return redirect()->back()->withErrors(['success','File upload successful']);
 
           }else{return redirect('profile')->withErrors(['Something went wrong!#FU101']);}//end if file->isValid
         }else{return redirect('profile')->withErrors(['Something went wrong! #FU102']);}//end if hasFile
@@ -110,20 +107,22 @@ class DocumentsController extends Controller
       if($file != null){
         if(Storage::exists($file->path)){
           if($file->member_id == auth()->user()->id){
-            //if the file belongs to the asking user
             if ($action == 'download'){
               $downloadName = $this->readableName($file).".".$file->extension;
               return Storage::download($file->path,$downloadName);
             }
             elseif ($action == 'delete'){
               $member = Member::find(auth()->user()->id);
-              if($file->type == 'application'){
+              if($member->misc->status !== 'incomplete')
+                return redirect('/profile')->withErrors(['warning','Action not allowed','You have already submitted an application.']);
+              $member = Member::find(auth()->user()->id);
+              /*if($file->type == 'application'){
                 $member->status = 'incomplete';
                 $member->save();
-              }
+              }*/
               $file->delete();
-              Storage::delete($file->path,$downloadName);
-              return redirect('profile');
+              Storage::delete($file->path);
+              return redirect()->back();
             }else return redirect('profile')->withErrors('Unsupported action!');
           }else return redirect('profile')->withErrors('Access denied!');
         }else return redirect('profile')->withErrors('File not found. #FA101');
@@ -149,5 +148,17 @@ class DocumentsController extends Controller
           return "Invalid file type";
           break;
       }
+    }
+
+    public static function docTypes()
+    {
+      return [
+        'ap' => 'application',
+        'id' => 'idpassport',
+        'pr' => 'proofop',
+        'ne' => 'nextofkin',
+        'be' => 'beneficiary',
+        'su' => 'supportingdoc',
+      ];
     }
 }
